+++ frontend/src/App.jsx
@@ -7,6 +7,8 @@ import AccountDetailPage from "./pages/accounts/AccountDetailPage.jsx";
 import PieDetailPage from "./pages/pies/PieDetailPage.jsx";
 import SearchPage from "./pages/search/SearchPage.jsx";
 import HoldingTickerPage from "./pages/holdings/HoldingTickerPage.jsx";
+import WatchlistsPage from "./pages/watchlists/WatchlistsPage.jsx";
+import GoalsPage from "./pages/goals/GoalsPage.jsx";
 import NotFoundPage from "./pages/errors/NotFoundPage.jsx";
 import TermsAndConditionsPage from "./pages/TermsAndConditionsPage.jsx";
 import PrivacyPolicyPage from "./pages/PrivacyPolicyPage.jsx";
@@ -31,6 +33,8 @@ function App() {
           <Route path="/accounts/:accountId/pies/:pieId" element={<PieDetailPage />} />
           <Route path="/search" element={<SearchPage />} />
           <Route path="/holdings/:ticker" element={<HoldingTickerPage />} />
+          <Route path="/watchlists" element={<WatchlistsPage />} />
+          <Route path="/goals" element={<GoalsPage />} />
           <Route path="*" element={<NotFoundPage />} />
         </Route>
       </Routes>